+++ lms/src/main/java/com/example/lms/services/UserService.java
@@ -5,8 +5,11 @@ import com.example.lms.dto.BookOutDTO;
 import com.example.lms.dto.UserOutDTO;
 import com.example.lms.dto.UserRegisterDTO;
 import com.example.lms.entity.Book;
+import com.example.lms.exception.ResourceAlreadyExistsException;
 import com.example.lms.mapper.UserMapper;
+import com.example.lms.repository.IssuanceRepository;
 import com.example.lms.repository.UserRespository;
+import com.example.lms.utils.PasswordGenerator;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
@@ -15,6 +18,7 @@ import org.springframework.core.env.Environment;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -23,8 +27,7 @@ import com.example.lms.entity.User;
 
 import javax.crypto.SecretKey;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 @Service
 public class UserService {
@@ -36,7 +39,13 @@ public class UserService {
     private UserMapper userMapper;
 
     @Autowired
+    private SMSService smsService;
+
+    @Autowired
     private PasswordEncoder passwordEncoder;
+
+    @Autowired
+    private IssuanceRepository issuanceRepository;
 
     @Autowired
     private Environment env;
@@ -45,7 +54,6 @@ public class UserService {
         User user = userRespository.findByMobileNumber(mobileNumber).orElseThrow(
                 () -> new UsernameNotFoundException("User not found for " + mobileNumber)
         );
-
 
         UserOutDTO userOutDTO = userMapper.toDTO(user);
         return userOutDTO;
@@ -62,9 +70,14 @@ public class UserService {
 
     public UserOutDTO deleteUserByMobile(String mobileNumber){
         User user = userRespository.findByMobileNumber(mobileNumber).orElseThrow(
-                () -> new UsernameNotFoundException("User not found for " + mobileNumber)
-        );;
+                () -> new UsernameNotFoundException("User details not found for this mobile number - " + mobileNumber)
+        );
 
+        boolean isBookIssued = issuanceRepository.existsByUserIdAndStatus(user.getId(), "Issued");
+        if(isBookIssued){
+            throw new IllegalStateException("This user has some books issued so it can't be deleted!");
+        }
+        issuanceRepository.deleteAllByUserIn(Collections.singletonList(user));
         userRespository.deleteById(user.getId());
         UserOutDTO userOutDTO = userMapper.toDTO(user);
         return userOutDTO;
@@ -84,12 +97,12 @@ public class UserService {
         return total;
     }
 
-    public Page<UserOutDTO> getUsersPaginated(int pageNumber, int pageSize, String search){
+    public Page<UserOutDTO> getUsersPaginated(int pageNumber, int pageSize, String sortBy, String sortDir, String search){
         Page<User> page;
         if(search!=null && !search.isEmpty()){
             page = userRespository.findByMobileNumberContainingIgnoreCaseAndRole(search,"ROLE_USER", PageRequest.of(pageNumber,pageSize));
         } else {
-            page = userRespository.findByRole( "ROLE_USER", PageRequest.of(pageNumber,pageSize));
+            page = userRespository.findByRole( "ROLE_USER", PageRequest.of(pageNumber, pageSize, Sort.Direction.fromString(sortDir), sortBy));
         }
         return page.map(user -> userMapper.toDTO(user));
     }
@@ -97,12 +110,38 @@ public class UserService {
     public UserOutDTO updateUserByMobile(String mobileNumber, UserRegisterDTO userRegisterDTO){
 
         User user = userRespository.findByMobileNumber(mobileNumber).orElseThrow(
-                () -> new UsernameNotFoundException("User not found for mobile no. " + mobileNumber)
-        );;
+                () -> new UsernameNotFoundException("User details not found for this mobile number - " + userRegisterDTO.getMobileNumber())
+        );
+
+        Optional<User> optionalUser = userRespository.findByMobileNumber(userRegisterDTO.getMobileNumber());
+        if (optionalUser.isPresent()) {
+            User thisUser = optionalUser.get();
+            if (thisUser.getId() != user.getId()) {
+                throw new ResourceAlreadyExistsException("User already exists for this mobile number - " + userRegisterDTO.getMobileNumber());
+            }
+        }
+
+        optionalUser = userRespository.findByEmail(userRegisterDTO.getEmail());
+        if (optionalUser.isPresent()) {
+            User thisUser = optionalUser.get();
+            if (thisUser.getId() != user.getId()) {
+                throw new ResourceAlreadyExistsException("User already exists for this email - " + userRegisterDTO.getEmail());
+            }
+        }
 
         user.setMobileNumber(userRegisterDTO.getMobileNumber());
         user.setName(userRegisterDTO.getName());
         user.setEmail(userRegisterDTO.getEmail());
+        if (userRegisterDTO.getPassword() != null && userRegisterDTO.getPassword().length() >= 3) {
+            String encodedPassword = userRegisterDTO.getPassword();
+            byte[] decodedBytes = Base64.getDecoder().decode(encodedPassword);
+            String decodedPassword = new String(decodedBytes);
+            userRegisterDTO.setPassword(decodedPassword);
+            user.setPassword(passwordEncoder.encode(userRegisterDTO.getPassword()));
+        }
+        if (user.getRole() == null || user.getRole().isEmpty()) {
+            user.setRole("ROLE_USER");
+        }
         User updated = userRespository.save(user);
 
         UserOutDTO userOutDTO = userMapper.toDTO(updated);
@@ -110,9 +149,37 @@ public class UserService {
     }
 
     public UserOutDTO createUser(UserRegisterDTO userRegisterDTO){
+        Optional<User> optionalUser = userRespository.findByMobileNumber(userRegisterDTO.getMobileNumber());
+        if(optionalUser.isPresent()){
+            throw new ResourceAlreadyExistsException("User with this mobile number already exist - " + userRegisterDTO.getMobileNumber());
+        }
+
+        optionalUser = userRespository.findByEmail(userRegisterDTO.getEmail());
+        if(optionalUser.isPresent()){
+            throw new ResourceAlreadyExistsException("User with this Email aleready exist - " + userRegisterDTO.getEmail());
+        }
+
         User user = userMapper.toEntity(userRegisterDTO);
-        user.setPassword(passwordEncoder.encode(userRegisterDTO.getPassword()));
+
+        String newPassword = PasswordGenerator.generatePassword(10);
+        user.setPassword(passwordEncoder.encode(newPassword));
+        if (user.getRole() == null || user.getRole().isEmpty()) {
+            user.setRole("ROLE_USER");
+        }
         User savedUser = userRespository.save(user);
+
+        String message = String.format( "\nWelcome %s\n" +
+                        "You have been successfully registered to Page Palace\n" +
+                        "Below are your login details\n" +
+                        "Username: %s (AND) %s\n" +
+                        "Password: %s",
+                savedUser.getName(),
+                savedUser.getMobileNumber(),
+                savedUser.getEmail(),
+                newPassword);
+
+        smsService.sendSms(savedUser.getMobileNumber(), message);
+
         UserOutDTO userOutDTO = userMapper.toDTO(savedUser);
         return userOutDTO;
     }
